Add fallback for data updater, which is needed b/c FWS data is spotty.

+++ data_updater.R
@@ -6,17 +6,29 @@ library(rvest)
 print(Sys.Date())
 
 cur <- readRDS("state_occ.rds")
-pg <- read_html("https://ecos.fws.gov/ecp/pullreports/catalog/species/report/species/export?format=htmltable&distinct=true&columns=%2Fspecies%40cn%2Csn%2Cstatus%2Cdesc%2Clisting_date%3B%2Fspecies%2Frange_state%40name%3B%2Fspecies%2Ffws_region%40desc%3B%2Fspecies%2Ftaxonomy%40group%3B%2Fspecies%2Frange_state%40abbrev&sort=%2Fspecies%40sn%20asc&filter=%2Fspecies%40status%20in%20('Endangered'%2C'Threatened')&filter=%2Fspecies%40country%20!%3D%20'Foreign'")
-new <- html_table(pg)[[1]]
+att1 <- try(
+  pg <- read_html("https://ecos.fws.gov/ecp/pullreports/catalog/species/report/species/export?format=htmltable&distinct=true&columns=%2fspecies%40cn%2csn%2cstatus%2cdesc%2clisting_date%3b%2fspecies%2frange_state%40name%3b%2fspecies%2ffws_region%40desc%3b%2fspecies%2ftaxonomy%40group%3b%2fspecies%2frange_state%40abbrev&sort=%2fspecies%40sn%20asc&filter=%2fspecies%40status%20in%20('endangered'%2c'threatened')&filter=%2fspecies%40country%20!%3d%20'foreign'")
+)
 
-if(nrow(new) > 2000) {
-  if(digest(cur) != digest(new)) {
-    file.rename("state_occ.rds", paste0("state_occ_", Sys.Date(), ".rds"))
-    saveRDS(new, "state_occ.rds")
-    print("Updated data saved.")
+if(class(att1) == "try-error") {
+  att2 <- try(
+    pg <- read_html("https://ecos.fws.gov/ecp/pullreports/catalog/species/report/species/export?format=htmltable&distinct=true&columns=%2fspecies%40cn%2csn%2cstatus%2cdesc%2clisting_date%3b%2fspecies%2frange_state%40name%3b%2fspecies%2ffws_region%40desc%3b%2fspecies%2ftaxonomy%40group%3b%2fspecies%2frange_state%40abbrev&sort=%2fspecies%40sn%20asc&filter=%2fspecies%40status%20in%20('endangered'%2c'threatened')&filter=%2fspecies%40country%20!%3d%20'foreign'")
+  )
+}
+
+if(exists("pg")) {
+  new <- html_table(pg)[[1]]
+  if(nrow(new) > 2000) {
+    if(digest(cur) != digest(new)) {
+      file.rename("state_occ.rds", paste0("state_occ_", Sys.Date(), ".rds"))
+      saveRDS(new, "state_occ.rds")
+      print("Updated data saved.")
+    } else {
+      print("No data changes.")
+    }
   } else {
-    print("No data changes.")
+    print("Something is amiss, data aren't big enough.")
   }
 } else {
-  print("Something is amiss.")
+  print("No data sent from FWS.")
 }
